refactor: Rename folder from InfoCard to WelcomeSection.
