+++ components/welcomeSection/InfoCard.tsx
